Skip optional variables which are set to None
Before, optional variables with a value of `None` were encoded as JSON `null` values.
This made it impossible to remove optional variables from the API, because the client
would still try to set a (null) value for that variable, which causes a Sangria error.

Now, optional variables are only encoded if they have a value. None values are removed
from the JSON altogether, making it possible to deprecated variables.