getResult method now takes implicit backend.
You now need to pass in an sttp backend to the getResult method. Because
the first type to the SttpBackend type is a higher kinded type, you
can't directly check the type and this is the way that works.

We could have had a separate method which uses the sync backend but this
feels more scala-y.