Upgrade sttp
The new backend check performance tests are using a newer version of
cats-effect which in turn needs a newer version of sttp.

It seems like a good time to upgrade this. There may be conflicts if we
try to use this somewhere that already has sttp elsewhere but we should
be able to fix those without too many issues.